Convert default output to new console output

+++ Source/Utilities/RandomNoise.h
@@ -11,13 +11,13 @@ public:
 
 	double costerp(double a, double b, double blend);
 
-	double createNoise(int x, int y, int z);
+	double createNoise(double x, double y, double z);
 
 	double noise(double x, double y, double z, int octaves, double roughness);
 
 	double noise(double x, double y, double z);
 
-	double getSmoothNoise(int x, int y, int z);
+	double getSmoothNoise(double x, double y, double z);
 
 	double getInterpolatedNoise(double x, double y, double z);
